% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTALYS.R
\name{initClusterTALYS}
\alias{initClusterTALYS}
\title{Setup Cluster for TALYS}
\usage{
initClusterTALYS(clust, talysExe, runOpts = NULL, calcsPerJob = 100)
}
\arguments{
\item{clust}{object as created by the package \code{clusterSSH}}
}
\value{
list with functions to perform TALYS calculations
}
\description{
Setup a computing cluster for TALYS calculations
}
